finaling pub sub playerdeath
finaling pub sub playerdeath
