+++ Assets/Scenes/Gameplay.unity
@@ -1576,7 +1576,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1332495725}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e8cce8976e0c5ae4fbb7d15be94b6bd8, type: 3}
+  m_Script: {fileID: 11500000, guid: b4810b11a891aab4cac27ece2ebf34da, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   leftKey: 97
@@ -1981,6 +1981,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   GameView: {fileID: 1431874240}
+  PlayerView: {fileID: 0}
 --- !u!4 &1627817140
 Transform:
   m_ObjectHideFlags: 0
