+++ Assets/Scripts/Gameplay/GpLauncher.cs
@@ -3,6 +3,7 @@ using Agate.MVC.Core;
 using Aja.Boot;
 using Aja.Game;
 using Aja.Score;
+using Aja.Player;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -13,6 +14,7 @@ namespace Aja.Gameplay
     {
         public override string SceneName => "Gameplay";
         private GameController _game;
+        private PlayerController _player;
 
         protected override IConnector[] GetSceneConnectors()
         {
@@ -29,12 +31,14 @@ namespace Aja.Gameplay
             {
                 new LivesController(),
                 new ScoreController(),
-                new GameController()
+                new GameController(),
+                new PlayerController()
             };
         }
 
         protected override IEnumerator InitSceneObject()
         {
+            _player.SetView(_view.PlayerView);
             _game.SetView(_view.GameView);
             yield return null;
         }
