+++ Assets/Scripts/Gameplay/GpView.cs
@@ -1,5 +1,6 @@
 using Agate.MVC.Base;
 using Aja.Game;
+using Aja.Player;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -9,5 +10,6 @@ namespace Aja.Gameplay
     public class GpView : BaseSceneView
     {
         public GameView GameView;
+        public PlayerView PlayerView;
     }
 }
