+++ Assets/Scripts/Gameplay/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1eb5ea8dfbaacc942b3cfbf619811a08
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
