+++ Assets/Scripts/Gameplay/Player/PlayerModel.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Aja.Player
+{
+    public class PlayerModel : MonoBehaviour
+    {
+
+    }
+}
