+++ Assets/Scripts/Gameplay/Player/PlayerModel.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 90876538033ab0e478de74974366c810
+guid: fd209901340941f4389af693ab2c2a5f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
