+++ Assets/Scripts/Gameplay/Player/PlayerView.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 90876538033ab0e478de74974366c810
+guid: b4810b11a891aab4cac27ece2ebf34da
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
