+++ Assets/Scripts/Gameplay/ScoreView.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 90876538033ab0e478de74974366c810
+guid: 14c9baa9885291e45ab525079a0e040d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
